backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
tensorflow==2.20.0
Pillow==11.0.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
gunicorn==23.0.0
